Overall : 	Warnings ..

+++ play_cphb/src/Chapter_3_insertion_sort.cpp
@@ -3,9 +3,9 @@
 int cphb::insertion_sort_v1( std::vector<int> &v )
 {
 	int inversions = 0;
-	for ( int i = 0 ; i < v.size()-1 ; i ++ )
+	for ( size_t i = 0 ; i < v.size()-1 ; i ++ )
 	{
-		for ( int j = i+1; j > 0; j -- )
+		for ( size_t j = i+1; j > 0; j -- )
 		{
 			if ( v[j] <= v[j-1] ) 
 			{
@@ -20,9 +20,9 @@ int cphb::insertion_sort_v1( std::vector<int> &v )
 int cphb::insertion_sort_v2( std::vector<int> &v )
 {
 	int inversions = 0;
-	for ( int i = 0 ; i < v.size()-1 ; i ++ )
+	for ( size_t i = 0 ; i < v.size()-1 ; i ++ )
 	{
-		for ( int j = i+1; j > 0; j -- )
+		for ( size_t j = i+1; j > 0; j -- )
 		{
 			if ( v[j] <= v[j-1] ) 
 			{
